Service worker proof of concept
