+++ example-app-router/app/ssr/page.tsx
@@ -1,6 +1,7 @@
 
-import { useAuth } from '@/auth/context'
-import styles from './page.module.css'
+import { headers } from 'next/headers'
+import styles from '../page.module.css'
+
 
 function Content() {
   return (
@@ -11,6 +12,9 @@ function Content() {
 }
 
 export default function SSR() {
+  const headersList = headers()
+  // console.log('headersList', headersList)
+  console.log('headersList.get("Authorization")', headersList.get("Authorization"))
   return (
     <main className={styles.main}>
       <Content />
