+++ example-app-router/package.json
@@ -19,5 +19,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "5.2.2"
+  },
+  "devDependencies": {
+    "workbox-webpack-plugin": "^7.0.0"
   }
 }
