+++ example-app-router/service-worker.ts
@@ -1,0 +1,4 @@
+
+import  { initializeAuthSW } from './sw'
+
+initializeAuthSW()
